+++ CHANGELOG.md
@@ -5,11 +5,40 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.10.0] - 2023-11-10
 
 ### Added
 
-- `format_solutes_dict()` method added into the utils module to help format solutes dictionaries with a unit.
+- Docs: new tutorial notebooks
+- Docs: remove duplicate contributing pages (Closes [#68](https://github.com/KingsburyLab/pyEQL/issues/68))
+- `Solution`: new method `to_file()` for more convenient saving Solution object to json or yaml files. (@kirill-push)
+- `Solution`: new method `from_file()` for more convenient loading Solution object from json or yaml files. (@kirill-push)
+- `Solution`: new classmethod `from_preset()` to `replace pyEQL.functions.autogenerate()` and instantiate a solution from a preset composition. (@kirill-push)
+
+### Fixed
+
+- Restored filter that suppresses duplicate log messages
+
+### Deprecated
+
+- `replace pyEQL.functions.autogenerate()` is now deprecated. Use `from_preset` instead.
+
+## [0.9.2] - 2023-11-07
+
+### Fixed
+
+- Restored filter that suppresses duplicate log messages
+
+## [0.9.1] - 2023-11-04
+
+### Added
+
+- `format_solutes_dict()` method added into the utils module to help format solutes dictionaries with a unit. (@DhruvDuseja)
+
+### Changed
+
+- Native property database is now instantiated on `pyEQL.__init__` to speed subsequent access by `Solution`
+- Numerous performance optimization increased the speed of many `Solution` property and method calls by 3x-10x
 
 ## [0.9.0] - 2023-10-17